# services/portal/internal/services/api/openapi/etr_study_app/fragments/schedule/getScheduleHistory.yaml
openapi: 3.1.0
info:
  title: Study App API · etr_study_app canon
  version: v1
  x-build-history: canon-first
  x-portal-spec: /services/portal/internal/services/api/reference/schedule/operations/get-api-v1-schedule-history.html
paths:
  /api/v1/schedule/history:
    get:
      operationId: getScheduleHistory
      x-implementation-status: implemented
      summary: Per-day review counts for the trailing N days (heat-map source).
      description: |
        Returns one bucket per UTC calendar day for the trailing `days`
        window, zero-filled — a day with no reviews is present with
        `count: 0`, so the client renders a gap-free grid. Each bucket
        also carries a pre-computed `intensity` (0–4) so heat-map cell
        shading needs no client-side maths.

        Read-only; no `Idempotency-Key`. Requires `X-API-Key`.
      tags: [Schedule]
      parameters:
        - $ref: "../../_shared/headers/ApiKey.yaml"
        - name: system_user_id
          in: query
          required: true
          description: External user id (composite key with `system_uuid`).
          schema:
            type: string
            minLength: 1
            maxLength: 36
          example: "learner-01"
        - name: system_uuid
          in: query
          required: true
          description: Source system UUID (composite key with `system_user_id`).
          schema:
            type: string
            format: uuid
          example: "b2c3d4e5-0002-4000-8000-000000000002"
        - name: days
          in: query
          required: false
          description: Size of the trailing window in days.
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 90
          example: 90
      responses:
        "200":
          description: OK — zero-filled per-day buckets, oldest first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduleHistoryResponse" }
              examples:
                three_days:
                  summary: Three-day window with one quiet day
                  value:
                    days:
                      - { date: "2026-07-17", count: 6, intensity: 3 }
                      - { date: "2026-07-18", count: 0, intensity: 0 }
                      - { date: "2026-07-19", count: 2, intensity: 1 }
                    computed_at: "2026-07-19T09:00:00+00:00"
        "401": { $ref: "../../_shared/responses/Error401.yaml" }
        "404":
          description: Not Found — the user does not exist for this tenant.
          content:
            application/json:
              schema: { $ref: "../../_shared/schemas/ApiErrorResponse.yaml" }
              examples:
                user_not_found:
                  summary: No user matches the composite key
                  value:
                    code: "USER_404"
                    key: "USER_NOT_FOUND"
                    message: "User with this `system_user_id` and `system_uuid` was not found."
                    source: "business"
        "422":
          description: Unprocessable Content — query parameters failed validation.
          content:
            application/json:
              schema: { $ref: "../../_shared/schemas/ValidationErrorResponse.yaml" }
              examples:
                days_out_of_range:
                  summary: "`days` outside 1–365"
                  value:
                    error_type: "validation_error"
                    endpoint: "GET /api/v1/schedule/history"
                    errors:
                      - code: "COMMON_000"
                        key: "COMMON_VALIDATION_ERROR"
                        message: "Query parameter `days` must be between 1 and 365."
                        field: "days"
                        source: "validation"
        "429": { $ref: "../../_shared/responses/Error429.yaml" }
components:
  schemas:
    ScheduleHistoryResponse:
      type: object
      required: [days, computed_at]
      properties:
        days:
          type: array
          description: "One bucket per UTC calendar day, oldest first, zero-filled."
          items:
            type: object
            required: [date, count, intensity]
            properties:
              date:
                type: string
                format: date
                description: "UTC calendar day this bucket covers."
              count:
                type: integer
                minimum: 0
                description: "Reviews logged on this day."
              intensity:
                type: integer
                minimum: 0
                maximum: 4
                description: "Bucketed density 0–4 for heat-map cell shading."
        computed_at:
          type: string
          format: date-time
          description: "Server timestamp when the window was computed (RFC 3339, UTC)."